# [configgrpc] Use own compressors for zstd